Player Pick Up Mechanics
This code is to pick up any GAMEOBJECT labeled "ITEM" you can change the
label to whatever the object is going to be in the future. Since I don't
know what we are going to be picking up there is no effect in picking up
anything yet. In addition the ITEM object will be destroyed on contact
with JUST THE PLAYER. Please insert this code in the player tab.