Player Pick Up mechanics
I forgot to put in the point system in the script so I just recently
added that. Other than that put this script in the player script. When
the player collides with a game object call "Item" (You can change it
later) it deletes the game object and collects 1 point