Patrolling Monster Vision Trigger Runs After Player
When the player triggers the line cast of the patrolling monster the
monster will run after the player until vision is broken or currently
until the patrol distance that is currently set. This could be changed
later in production. The speed is increased by a multiplier that can be
set in the inspector with current default of 2.0f