Title screen (prompt + 2 buttons)
Here are what I had in the scene
- An animated background. Just a random texture.
- I lifted the logo from the art team's Trello
- Some sort of menu "flow", from the Press Start prompt to the buttons
then fade out to the test level.
- Custom button that works via a BoxCollider2D. I'm still not familiar
with Canvas buttons so I don't know how flexible they are, but with
these you can dress them up however you want.
- The buttons are animated sprites that can change states. Just trying
to familiarize myself with the Animator.
- Sound effects for good measure, although they're just there and not
attached to anything for now

Everything in the img and sfx folders are placeholders I cooked up
myself so there's something to display to the screen. They need to be
replaced by the real stuff later.

script/MainTitle.cs
This contains the code for the menu. It runs in a giant IEnumerator that
advances depending on the user's action. It is a bit convoluted, but the
important functions are:
_Main(): where everything happens
MouseCheck(), KeyCheck(): default to mouse; the key check works but is
commented out for now
StartGame(), Quit(): loads the scene "TestScene", exit out of the
application

script/CustomButton.cs
Script for the 'Test' and 'Quit' button.

script/fx____.cs
Experimental scripts to make things pretty. They're not a key part of
the title screen.

Errors
- Every once in a while a couple NullReferenceExceptions are thrown on
the animation states. Not sure what's causing it.