Vision Less Patrolling Enemy
A monster will spawn and patrol a certain distance before despawning. I
created a spawner prefab which turns invisible and won't react or
trigger with the player during game time. It will spawn vision less
patrolling enemies that will move in one direction and will kill the
player on contact. The direction of the spawner can be set in the
inspector and will spawn monsters going to the left or right depending
on whether "Facing Right" is selected or not.