Some Changes
This is the Ceiling monster code some changes appears to take place when
running the ceiling monster with the code so take heed of these changes.
For LINE RENDERER in the inspector the element 1 should be X)0 Y)-1 Z)0
and everything else should be the same as the patrolling monster for
just the line renderer as for the box collider the IS TRIGGER is
checked, and for my code please be weary of the movement and speed those
can be easily fixed in my code by changing the number of how has you
want to go they are public variables labeled movement and speed. As for
the Time counter for when the monster is stunned you can change it to 3
seconds the unity keeps changing it to 30. The Line Cast Distance is 20.
Other than that just email me if you have any questions.