Fade only happens when player is in room
So these scripts aren't as condensed but they are just repetitive codes
from the others.

For this to work
we need to create individual canvases that has the fading Darkness
script, image component, and canvas group component. The amount of
canvases depends on how many rooms we need. With that said, each
canvases are tagged with "Canvas#" the # is just any number. For the
flickering lights, you dont need to create a canvas, just put the room
detection collider. This segways into the RoomDetection Colliders. You
create an empty child with only a box collider that acts as a trigger.
They don't need a tag but each RoomDetection Colliders need to be named
properly. The name is RoomDetection# same as before the # is just any
number corresponding the room. However, you do need this in the
flickering light since when you pass each room, the code is set so that
when you pass, the alpha in the previous room turns 0.

Confusing, i know, but it works, if you are confused ask me.