Music Mute/Slider + Brightness Slider
Main menu had OPTIONS button added containing a back button, mute button
for music, and slider controls for both the volume and the brightness of
the game.

Incidentally due to this a canvas has been added for the options menu
which was another task altogether but was not assigned to us. The lack
of a sound file did not help test the volume slider and mute button and
so a filler music file was added to the project to play in the
background of the main menu and the test scene.

Glitches:
The brightness and volume slider so far only adjust the values for the
main menu and so they will be refined.

The music for the game also overlaps itself when brought back to the
main menu and will be looked at as well.

The Values for the Music Volume and the Brightness has not yet been
prepped to carry over to the Test Scene and may take some more
attention.