Candidate Spotted Indicator Repair - Moved activating the SpottedIndicator to LateUpdate, but left deactivating it in Update.
After numerous trial and errors it seems that having two .SetActive statements in one of the update methods won't work as expected even if it is logically correct. Seperating it into Update and LateUpdate fixes the problem and it is now consistence starting the game fresh or restarting in game.