Options Menu with Sliders
Added sliders to the options menu, plus a temporary sprite to test the
brightness change.

Brightness is changed in code via adjusting the ambient light's RGB
values.
A sprite needs its Material set to Diffuse to receive the effect.

The background had its Material set to Diffuse, drag the Brightness
slider to change its visibility.
Adjust the min, max, default values of the slider if needed.

All changes only happen within this scene. The light in the test scene
also needs to be updated with the correct value and sprites need to be
set to Diffuse.

BGM and SE sliders aren't working yet.