Double Double Hallucination Room
We changed some variables to public in the player player control.
Everything else is set in the Prefab. The double double code has some
commented out sprites for later use in adding more details in the
hallucination room. In addition, when the player collides with the
double image then the image will disappear. We will add special affect
next week. Brought to you by the Chao and Tol production (Alex and
Jimmy).