Fixed Patrol spawner+monster; Chaser with periodic pause
For the Patrol monster and spawner:
MonsterRoom3.prefab is the room with both the fixed Patrol and Patrol
Spawner. If you want it to move left, go to the spawner and uncheck
FacingRight. To switch out the old room:
MapManager -> Random Map Generation -> Reg Room -> change Element
3/Element 10 to this room

For the Chaser:
Just replace its script with the new script.