FixedFogRoom
So there is  a lot to commit since I am using some textures and shaders
imported from unity. The standard assets - i dont know whats wrong with
it but in it ... It consists of the particle steam material (idk about
particle steam mobile), shaders. alpha blend, and black and white
clouds.