Earthquake Room
Made a script and room prefab. Simply put, the script makes the camera
move to a random location within a unit sphere(z axis locked to original
camera z point) while the player is within the bounds of the room.  When
the player leaves the room the camera goes back to its original
position.  Uses a public variable to increase or decrease the intensity
of the shaking.