Bug Hallucination Room
These are the assets for the bug hallucination room. When the player
enters the room bugs will spawn and move randomly around the room until
the player leaves. More bugs mean a cooler effect but will decrease
performance. There are bugs breaking the 4th wall by crawling on the
camera.