modify chasing monster to patrol the entire length of the room it is in.
I just commented out things related to patrol distance since that was
being used to cap the distance in the ChasingMonster script. And I made
a new script that used an OnTriggerExit2D function that checked if the
monster was a ChasingMonster and then reference the FlipEnemy() method
in the ChasingMonster script. The Chasing Monster needs a rigidbody with
gravity set to zero but i'm not sure how to commit that. Sorry.