Final Level Improvements
Here are the improvements to the final level!
When the player enters the long hall the cutscene will initiate. The
player will be unable to move and unable to go back up into the
attic(visual cue could be added later to tell the player they can't go
back) When the player reaches the end and uses the door to get to the
basement there will be an item when the player uses it the player will
freeze and load the credits.