Cut to black when dead
I just took out the timer in the update method. and made it so that when
dead the overlay is set to active. There are some dead code but i dont
want to tinker with it without knowing that it will affect other things.
oh and I dont see any meta files that comes with this script