Ken Miller -  Ending Cutscene
When the player uses the item in the basement the cutscene will
activate. The player will freeze and the hunter monster will spawn
outside and move into the room. When the monster is inside the room the
monster will fade and the silhouette(the prefab is temporarily a black
player model, to be changed later once art assets are made for the
brother) will appear. Then the silhouette will disappear and the screen
will turn white and the credits will start rolling.

Also in the commit:
-Fixed the doors leading from the attic to the long hall and from the
long hall to the basement. The sprites were not set in the inspector so
I made the corrections.
-Moved the hunter monster up by 10 so that it is not clipping through
the ground.