Level Complete Cutscene
The level complete cutscene is implemented. The camera centers on the
items, the glow fades away, camera pans back to the player/bed/light
bulb section of the room, the light bulb glows, camera goes black with
light bulb and glow still visible, then finally the light bulb and glow
fades out and the next level is loaded.

Additional changes needed:
The MapManager prefab for the AtticLevel was updated.
-Tags were added to the item glow object to make searching for them to
put into an array easier.
-Light bulb glow name was changed to make searching easier
-A black out canvas was added to make the camera black out during the
cutscene

Additional tags implemented
-ParticleGlow tag included to make things easier
-"AboveUI" layout used to make the light bulb and light bulb glow stay
visible when the camera blacks out.

Emissions rate for the glow particles increased to 6 to make fade in/out
less choppy.
Color was changed in respect with the emissions rate change to make the
particle look very similar to the original particle effect.