Speed up TV monster
For some reason no matter how much I changed the speedChasing variable,
the TV monster wouldn't go faster. So what I did was add a bool argument
in the MoveTurn() function, and set it in the if else statement where
MoveTurn() is called, since that statement already checks for if player
is seen. If true it would speed up the monster based on a constant that
i added to the movement variable. if it was false it would move at the
normal speed. It is set so that you can't outrun it, but it wont catch
you unless you stop moving. It basically chases you out of the room or
you have to hide.