Improve 4th Wall Bug
In all applicable scenes the camera bug must be attached to the camera
for it to follow and stay inside the camera. I believe I have covered
all the scenes necessary but if I have missed any just attach a camera
bug to the camera.

I moved the for loop to search for the bug from update to start as this
is more efficient if there is only going to be one camera bug. If in the
future more is desired just de-comment the for loop in update and attach
any new camera bugs to the camera.

I increased the degree of which the bugs change direction to give them
more of a random movement rather than before where they changed
direction only a tiny bit.