Slightly more polished main menu
Finished:

- Changed the background to imitate what's on the .gif. Graphics are in
the BG object.

- Added fade in/fade out. The fade image is a child of CONTROL.

- PlayerPrefs values should now be stored and initialized properly.

- PlayerPrefs reading/writing is now managed by the screens (i.e. the
audio screen contains code to initialize and set the volumes, etc.) The
reason for this is the sliders need to be set too, and sliders are only
accessible via the screens that own them. And as a result the code in
AudioHandlerScript.Awake() has been commented out.

- All buttons now have bigger hitboxes.

- Now the pause screen's time freezing is breaking the splash screen
instead, so I just fixed that script directly.

---

Notes:

- There was some audio handler initialization code in the menu control,
I moved it into the audio handler itself. It seems to be working.

- Bulb swing was disabled. Tick the script in the Bulb object if you
want the behavior back (though this also requires changes to the shadow
behavior thanks to ~physics~).

---

Todo:
- Add the remaining buttons
- Fix the texts on buttons
- Buttons: some kind of confirmation when pressed
- Sliders: bells+whistles
- Brightness->Preview panel: bells+whistles, use a better picture