widen the side linecast of the ceiling monster
So i was able to widen them but not in till i commented out the public
gameObjects endCastObj, leftCastObj, and rightCastObj (not sure what
they were suppose to do). I just replaced them in the update() and
fixedupdate() with the vectors endCast, rightCast, and leftCast because
in start() it seemed that they did all the work that i needed. After
that i was able to widen the side line with the detectionWidth in
start(). Let me know if you want me to change anything